add ability to cis_publisher.Publish object to return data from CIS Person API in various forms without actually publishing. This is also used internally by the Publish object while publishing
call Publish.get_known_cis_users() from cis_publisher.hris in order to
be able to match emails to user_ids using real user ids.

This will emit a critical log BUT NOT RAISE if a user exists in the HRIS
database but not the CIS database. This can happen in 2 cases:
- A terrible bug
- User has not been added to LDAP *yet* (most likely)