Update identities in publisher rules to be more scoped
This set of changes realigns the permissions granted to LDAP and
DinoPark with the capabilities present and used in production today.
(For example, DinoPark has no UI to modify SSH or PGP keys.)

These were prepared in #523 to allow LDAP to update LDAP keys that can't
be set by any other CIS integration, and this commit contains that work.