dockerfile: Merge dependency install commands
While it may be easier for diffs to have them on separate lines, our CI
runs are approaching fifteen minutes in GitHub Actions and a lot of that
time is spent running the pacman and other commands repeatedly. Each of
these package managers support multiple packages at once, so we take
advantage of that to gain a lot of wallclock efficiency back.