person/change Custom domains are managed elsewhere
Now that we're onboarding to Fastly WAF, this custom domain configuration will
need to be moved elsewhere.

For now, they're "clickops"'d to point to the Fastly endpoint, for lack of a
better place.

Previous records are:

    person.api.dev.sso.allizom.org CNAME dkvp8o2j3an0s.cloudfront.net
    change.api.dev.sso.allizom.org CNAME dnvuz7rflzcix.cloudfront.net

    person.api.test.sso.allizom.org CNAME d1xhywfw7lzv8t.cloudfront.net
    change.api.test.sso.allizom.org CNAME dqogv2ah2qsgb.cloudfront.net

Prod isn't set up quite yet, but for completion, the records are:

    person.api.sso.mozilla.com CNAME d2ot71lo28jdym.cloudfront.net
    change.api.sso.mozilla.com CNAME d1dx0nvrqzhzid.cloudfront.net

These were all updated to be `CNAME mozilla.map.fastly.net`.

Jira: IAM-1629, OPST-1768