build Upload AWS Lambda Layer
We update various places in SSM with a reference to the layer we upload. Today
(er, or rather, yesterday) those are:

    * `/iam/cis/development/lambda_layer_arn`
    * `/iam/cis/testing/lambda_layer_arn`
    * `/iam/cis/production/lambda_layer_arn`

I'll be keeping these around as a reference and _will not be removing them_, at
least until we fully cut over to the new SSM parameters.

For reference, the current respective values are:

* `arn:aws:lambda:us-west-2:320464205386:layer:cis_development:601`,
* `arn:aws:lambda:us-west-2:320464205386:layer:cis_testing:383`, and
* `arn:aws:lambda:us-west-2:320464205386:layer:cis_production:71`.

Jira: IAM-1669