Use DAX in the profile retrieval service
Specifically for v2UsersByAny, a query which is notoriously slow. To enable
using Dax, add a `&cached=true` request parameter.

AWS provides the amazon-dax-client library, which is similar to boto3. It has
some limitations, but those don't seem to affect how we're going to use it.

    The Amazon DAX client does not support table operations. Any table
    manipulation operations must use the regular Boto3 or botocore DynamoDB
    client.

    Paginators are not currently supported for DAX.

We'll probably want to think hard about if we actually use this. I'm not sure
which expectations/assumptions will break by returning cached results (if there
are any).

Source: https://pypi.org/project/amazon-dax-client/

Jira: IAM-1668