fix(identity-vault): pagination while scanning overrides in the API
In `parallel_dynamo.py`, we were paginating over the results for our callers.
While this simplifies the callers' code, it also increases their run time,
which is an issue when paired with AWS API Gateway.

The fix here is to do what we say we're going to do: allow the users to
paginate if they feel. We provide a way for users to specify where their next
page should start, via `scan`'s `exclusive_start_key` argument. We hint as much
in the return, naming this key `nextPage`.

But, doing this requires fixing our segmentation logic! Each segment receives
it's own `ExclusiveStartKey`, which needs to be used the next time we call our
next scan operation.

Jira: IAM-1793